move utility funcs in run_vllm_api_server.py to utils/
